% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_New_SiteID.R
\name{Get_New_SiteID}
\alias{Get_New_SiteID}
\title{Get_New_SiteID}
\usage{
Get_New_SiteID(x, CRDB)
}
\arguments{
\item{x}{Site ID or CRDB that you would like updated}

\item{CRDB}{TRUE = CRDB; FALSE = SiteID}
}
\value{
A Vector containing the updated Site ID/CRDB
}
\description{
This function updates an old Site ID or CRDB to the current Site ID/CRDB. The Fields names must remain unchanged and the CRDB field will need to be added to both HIER and OLD objects if CRDB = TRUE.
}
\examples{
Get_New_SiteID(x = "0000340", CRDB = FALSE)
[1] "08"

}
